S08P12C202-250 [BE] feat : 임베디드에서 받은 데이터 저장

+++ backend/src/main/java/com/ssafy/plant/dto/LikingDTO.java
@@ -6,11 +6,11 @@ import lombok.*;
 @NoArgsConstructor
 @ToString
 public class LikingDTO {
-    private int serialNo;
+    private long serialNo;
     private int liking;
 
     @Builder
-    public LikingDTO(int serialNo, int liking) {
+    public LikingDTO(long serialNo, int liking) {
         this.serialNo = serialNo;
         this.liking = liking;
     }
